Add setup for link, runtime-link, and toolset version.
This adds setting up b2 link as specified in package shared option (if present). It also sets up runtime-link, and toolset version as specified in the build settings. Not that for MSVC and Boost 1.64 one must run in the VS command shell for VS 14.1 (aka 15) to be correctly detected and used.